+++ apps/web/app/api/v1/client/[environmentId]/responses/[responseId]/route.ts
@@ -8,11 +8,9 @@ import { withV1ApiWrapper } from "@/app/lib/api/with-api-logging";
 import { sendToPipeline } from "@/app/lib/pipelines";
 import { getResponse } from "@/lib/response/service";
 import { getSurvey } from "@/lib/survey/service";
+import { formatValidationErrors } from "@/modules/api/lib/validation";
 import { validateOtherOptionLengthForMultipleChoice } from "@/modules/api/v2/lib/element";
-import {
-  formatValidationErrorsForV1Api,
-  validateResponseData,
-} from "@/modules/api/v2/management/responses/lib/validation";
+import { validateResponseData } from "@/modules/api/v2/management/responses/lib/validation";
 import { createQuotaFullObject } from "@/modules/ee/quotas/lib/helpers";
 import { validateFileUploads } from "@/modules/storage/utils";
 import { updateResponseWithQuotaEvaluation } from "./lib/response";
@@ -131,7 +129,7 @@ export const PUT = withV1ApiWrapper({
         return {
           response: responses.badRequestResponse(
             "Response validation failed",
-            formatValidationErrorsForV1Api(validationErrors),
+            formatValidationErrors(validationErrors),
             true
           ),
         };
