+++ apps/web/app/api/v1/management/responses/[responseId]/route.ts
@@ -8,10 +8,8 @@ import { TApiAuditLog, TApiKeyAuthentication, withV1ApiWrapper } from "@/app/lib
 import { sendToPipeline } from "@/app/lib/pipelines";
 import { deleteResponse, getResponse } from "@/lib/response/service";
 import { getSurvey } from "@/lib/survey/service";
-import {
-  formatValidationErrorsForV1Api,
-  validateResponseData,
-} from "@/modules/api/v2/management/responses/lib/validation";
+import { formatValidationErrors } from "@/modules/api/lib/validation";
+import { validateResponseData } from "@/modules/api/v2/management/responses/lib/validation";
 import { hasPermission } from "@/modules/organization/settings/api-keys/lib/utils";
 import { validateFileUploads } from "@/modules/storage/utils";
 import { updateResponseWithQuotaEvaluation } from "./lib/response";
@@ -156,7 +154,7 @@ export const PUT = withV1ApiWrapper({
         return {
           response: responses.badRequestResponse(
             "Validation failed",
-            formatValidationErrorsForV1Api(validationErrors),
+            formatValidationErrors(validationErrors),
             true
           ),
         };
