+++ apps/web/app/api/v2/client/[environmentId]/responses/route.ts
@@ -11,11 +11,9 @@ import { sendToPipeline } from "@/app/lib/pipelines";
 import { getSurvey } from "@/lib/survey/service";
 import { getElementsFromBlocks } from "@/lib/survey/utils";
 import { getClientIpFromHeaders } from "@/lib/utils/client-ip";
+import { formatValidationErrors } from "@/modules/api/lib/validation";
 import { validateOtherOptionLengthForMultipleChoice } from "@/modules/api/v2/lib/element";
-import {
-  formatValidationErrorsForV1Api,
-  validateResponseData,
-} from "@/modules/api/v2/management/responses/lib/validation";
+import { validateResponseData } from "@/modules/api/v2/management/responses/lib/validation";
 import { getIsContactsEnabled } from "@/modules/ee/license-check/lib/utils";
 import { createQuotaFullObject } from "@/modules/ee/quotas/lib/helpers";
 import { createResponseWithQuotaEvaluation } from "./lib/response";
@@ -121,7 +119,7 @@ export const POST = async (request: Request, context: Context): Promise<Response
   if (validationErrors) {
     return responses.badRequestResponse(
       "Response validation failed",
-      formatValidationErrorsForV1Api(validationErrors),
+      formatValidationErrors(validationErrors),
       true
     );
   }
